refactor to centralize functionality
